better management of the orange vs orange-canvas-base
